pyproject.toml
# pyproject.toml for the klab project

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "klab"
version = "0.1.0"
authors = [
  { name="Juan E Villegas", email="jev1@nyu.edu" },
]
description = "A KLayout package for controlling measurement instruments."
readme = "docs/README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Framework :: KLayout",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering",
]

# Core dependencies required for the project to run.
# These mirror the dependencies checked in the install.py script.
dependencies = [
    "pyvisa>=1.11.0",
    "pyyaml>=6.0",
    "numpy>=1.19.0"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "ruff",
]

[tool.setuptools]
packages = { find = {} }
package-dir = {"" = "klayout_package/python"}

[tool.setuptools.package-data]
klab = ["tech/*.yml"]

[project.urls]
"Homepage" = "https://github.com/jevillegasd/klab" # 
"Bug Tracker" = "https://github.com/jevillegasd/klab/issues" # 